# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Exemplo",
        "description": "Documentação da API de Exemplo",
        "contact": {
            "email": "contato@exemplo.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/livros": {
            "get": {
                "tags": [
                    "Livros"
                ],
                "summary": "Lista todos os livros",
                "description": "Retorna a lista de todos os livros",
                "operationId": "c2638f778b25c161b5b8189172216b67",
                "responses": {
                    "200": {
                        "description": "Lista de livros obtida com sucesso"
                    }
                }
            },
            "post": {
                "tags": [
                    "Livros"
                ],
                "summary": "Cria um novo livro",
                "description": "Cria um novo livro e vincula autores e assuntos",
                "operationId": "052dc25bf67cc3710e476ca8c1f49397",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "titulo",
                                    "editora",
                                    "edicao",
                                    "anoPublicacao",
                                    "valor",
                                    "autor",
                                    "assunto"
                                ],
                                "properties": {
                                    "titulo": {
                                        "type": "string",
                                        "example": "Livro Exemplo"
                                    },
                                    "editora": {
                                        "type": "string",
                                        "example": "Editora Exemplo"
                                    },
                                    "edicao": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "anoPublicacao": {
                                        "type": "string",
                                        "example": "2024"
                                    },
                                    "valor": {
                                        "type": "number",
                                        "format": "float",
                                        "example": "59.90"
                                    },
                                    "autor": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "assunto": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Livro criado com sucesso"
                    },
                    "400": {
                        "description": "Erro de validação"
                    }
                }
            }
        },
        "/livros/{id}": {
            "put": {
                "tags": [
                    "Livros"
                ],
                "summary": "Atualiza um livro",
                "description": "Atualiza as informações de um livro existente",
                "operationId": "34f924b16d17d726283baad63b144290",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do livro",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "titulo",
                                    "editora",
                                    "edicao",
                                    "anoPublicacao",
                                    "valor",
                                    "autor",
                                    "assunto"
                                ],
                                "properties": {
                                    "titulo": {
                                        "type": "string",
                                        "example": "Livro Atualizado"
                                    },
                                    "editora": {
                                        "type": "string",
                                        "example": "Editora Atualizada"
                                    },
                                    "edicao": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "anoPublicacao": {
                                        "type": "string",
                                        "example": "2024"
                                    },
                                    "valor": {
                                        "type": "number",
                                        "format": "float",
                                        "example": "79.90"
                                    },
                                    "autor": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "assunto": {
                                        "type": "integer",
                                        "example": 2
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Livro atualizado com sucesso"
                    },
                    "400": {
                        "description": "Erro de validação"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Livros"
                ],
                "summary": "Exclui um livro",
                "description": "Exclui logicamente um livro (status = 0) e atualiza seus vínculos",
                "operationId": "20ff1e31d004d5d38f9e8c62a3568b1f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do livro",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Livro excluído com sucesso"
                    },
                    "400": {
                        "description": "Erro ao tentar excluir"
                    }
                }
            }
        },
...
    },
...
}